pyproject.toml
[tool.poetry]
name = "currencies"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"
celery = "^5.3.4"
django-redis = "^5.4.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
filterwarnings = ["ignore::DeprecationWarning"]